Clean up the way the theme does linking by default
First, use .LinkTitle rather than .Title where appropriate.

Second, don't special case a hard-coded "About" page. Instead, the
links on the top right will go to pages that aren't in any section.

Third, page lists include pages that are in a section. This includes
pages in a given taxonomy, so custom content types work correctly.

Fourth, don't special case on "hidden" in the front-matter. This is no
longer needed because we handle exclusions by section. To hide content
you don't wish to have published, use the standard drafts functionality.

Fifth, retain the "about" parameter for the link on the avatar. By
deafult, this link will point to the home page. But users of the theme
can point it to a custom page (for example, a page without a section).